+++ package/CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.11-preview] - 2018-6-20
+### Added
+- Fix Spriteshape does not update when Sprites are reimported.
+- Fix SpriteShapeController in Scene view shows a different sprite when user reapplies a Sprite import settings
+- Fix Editor Crashed when user adjusts the "Bevel Cutoff" value
+- Fix Crash when changing Spline Control Points for a Sprite Shape Controller in debug Inspector
+- Fix SpriteShape generation when End-points are Broken.
+- Fix cases where the UV continuity is broken even when the Control point is continous.
+
 ## [1.0.10-preview] - 2018-4-12
+### Added
 - Version number format changed to -preview
 
 ## [0.1.0] - 2017-11-20